Extract common presentational- and base-component tests
