+++ src/components/navigation/test/LinkBase-test.js
@@ -1,5 +1,6 @@
 import { mount } from 'enzyme';
-import { createRef } from 'preact';
+
+import { testBaseComponent } from '../../test/common-tests';
 
 import LinkBase from '../LinkBase';
 
@@ -13,29 +14,7 @@ describe('LinkBase', () => {
   };
 
   // Tests common to base components
-  it('applies `ref` via `elementRef`', () => {
-    const elementRef = createRef();
-    createComponent({ elementRef });
-
-    assert.instanceOf(elementRef.current, Node);
-  });
-
-  it('applies supplied `className`', () => {
-    const wrapper = createComponent({ className: 'testClass' });
-
-    const allClasses = [...wrapper.childAt(0).getDOMNode().classList];
-
-    assert.deepEqual(allClasses, ['testClass']);
-  });
-
-  it('applies HTML attributes to outer element', () => {
-    const wrapperOuterEl = createComponent({ 'data-testid': 'foo-container' })
-      .childAt(0)
-      .getDOMNode();
-
-    assert.isTrue(wrapperOuterEl.hasAttribute('data-testid'));
-    assert.equal(wrapperOuterEl.getAttribute('data-testid'), 'foo-container');
-  });
+  testBaseComponent(LinkBase);
 
   // Tests specific to this base component
   it('adds common `rel` attributes', () => {
